build: Switch from eslint and prettier to Biome
Biome comes with a linter and a formatter. It's noticably faster than
eslint and prettier (although this isn't that important if you use an
editor that integrates those tools). Biome also has a decent level of
compatiblity with those prettier and eslint. It covers over 95 % of
prettier's test cases. Not sure about eslint but it does offer a
migration for it... So let's try it out and see how it goes.

Although we only switch on the main project. The templates stay on
eslint and prettier since those are more widely used and upstream GNOME
also uses eslint.