chore: check compliance with Conventional Commits via git hook
At first I tried to use @commitlint/cli but the speed wasn't that good.
So I decided to use a simple git hook to check the (header of the)
commit message for compliance with the Conventional Commits spec since
that's the most important part and only warn if the spec isn't followed.
The full compliance will be checked in the CI pipeline server-side since
speed isn't that important there.