test: add basic tests
I've used vitest since I started using Vue/Vite at work now, so getting
farmiliar with vitest seemed like a good idea. There is no other particular
reason for vitest. It certainly is helpful that it is fairly API compatible
with jest, so it's easy to google for help.

The tests only cover the core logic of processing the arguments since the
rest is basically just copying/writing files around based on the processed
arguments. So that should hopefully be straightforward enough... to not
bother with tests.

To test stuff I split up the main index.js file so that the modules can
be mocked.

Close: https://github.com/Leleat/create-gnome-extension/issues/21
Part-of: https://github.com/Leleat/create-gnome-extension/pull/42