fix: don't skip installing devDependencies in build script
1583024 moved the dependencies to devDependencies because extensions
don't need them at runtime. But the build script was skipping the install
of devDependencies when building the extension package if there was no
node_modules directory. This commit removes the --omit=dev flag from the
npm install command in the build script to fix that.

Part-of: https://github.com/Leleat/create-gnome-extension/pull/43