fix: change "mandatory" postinstall to optional hookup script
Apparently, the postinstall script to change the git hooks path causes an
error. For some reason I can't reproduce this on my linux machine, but
I can reproduce it on my MacBook... I can't be bothered to investigate
this further, so I'm just going to make it optional.

Part-of: https://github.com/Leleat/create-gnome-extension/pull/50